Mostly done fixing errors

+++ BusinessLayer/Models/CommentDisplay.cs
@@ -1,4 +1,5 @@
 using System;
+using BusinessLayer.Services;
 
 namespace BusinessLayer.Models
 {
@@ -6,35 +7,29 @@ namespace BusinessLayer.Models
     {
         // Currently logged-in user ID for comparison
         private static readonly uint CurrentUserId = ForumService.GetForumServiceInstance().GetCurrentUserId();
-        
         // Original comment data
         public ForumComment Comment { get; private set; }
-        
         // User information
         public User Author { get; private set; }
-        
         // Properties for easy binding
         public uint Id => Comment.Id;
         public string Body => Comment.Body;
         public int Score => Comment.Score;
         public string TimeStamp => Comment.TimeStamp;
         public uint AuthorId => Comment.AuthorId;
-        
         // User properties
         public string Username => Author.Username;
         public string ProfilePicturePath => Author.ProfilePicturePath;
-        
         // Indicates if the comment belongs to the current user (for delete button visibility)
         public bool IsCurrentUser => AuthorId == CurrentUserId;
-        
         // Create a CommentDisplay from a ForumComment
         public static CommentDisplay FromComment(ForumComment comment)
         {
             return new CommentDisplay
             {
                 Comment = comment,
-                Author = User.GetUserById(comment.AuthorId)
+                Author = User.GetUserById((int)comment.AuthorId)
             };
         }
     }
-} 
+}
